Rename directory name from flask to application
